o11y: move observability-related code to own crate
