+++ node/Cargo.toml
@@ -10,6 +10,7 @@ path = "src/lib.rs"
 [dependencies]
 anyhow = "1.0.66"
 apibara-core = { path = "../core" }
+apibara-observability = { path = "../observability" }
 arrayvec = "0.7.2"
 async-trait = "0.1.57"
 byte-unit = "4.0.14"
