+++ observability/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "apibara-observability"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+opentelemetry = { version = "0.18.0", features = ["trace", "metrics", "rt-tokio"] }
+opentelemetry-otlp = { version = "0.11.0", features = ["trace", "metrics", "grpc-tonic"] }
+thiserror = "1.0.32"
+tracing = "0.1.36"
+tracing-opentelemetry = "0.18.0"
+tracing-subscriber = { version = "0.3.15", features = ["std", "env-filter"] }
+tracing-tree = "0.2.2"
