+++ observability/README.md
@@ -1,0 +1,8 @@
+# Apibara Observability
+
+This crate contains helpers to setup observability with `tracing` and `opentelemetry`.
+
+
+## Usage
+
+Invoke the `init_opentelemetry` function at the beginning of your program.
